requirements.txt
pip
python==3.10.13
pytorch==2.0.1
torchvision==0.15.2
pytorch-cuda==11.8
pandas==1.5.2
joblib==1.1.1
transformers==4.32.1
torchmetrics==1.2.0
biopython==1.79
torchdata==0.4.1
wandb==0.15.11
sacremoses==0.0.53
pynvml==11.5.0
torchmetrics==1.2.0
torcheval==0.0.7
wget==3.2
azureml-mlflow==1.53.0
loralib==0.1.2
tensorboard==2.15.1
obonet==1.0.0
blosum==2.0.2